have a separate assign_costs rule
