+++ scripts/add_electricity.py
@@ -92,6 +92,12 @@ from _helpers import (
     sanitize_locations,
     update_p_nom_max,
 )
+from assign_costs import (
+    attach_dc_costs,
+    calculate_renewable_capital_cost,
+    update_electricity_costs,
+    update_transmission_costs,
+)
 
 idx = pd.IndexSlice
 
@@ -241,410 +247,6 @@ def attach_load(n, demand_profiles):
     demand_df = read_csv_nafix(demand_profiles, index_col=0, parse_dates=True)
 
     n.madd("Load", demand_df.columns, bus=demand_df.columns, p_set=demand_df)
-
-
-def attach_dc_costs(lines_or_links, costs, length_factor=1.0, simple_hvdc_costs=False):
-    if lines_or_links.empty:
-        return
-
-    if lines_or_links.loc[lines_or_links.carrier == "DC"].empty:
-        return
-
-    dc_b = lines_or_links.carrier == "DC"
-    if simple_hvdc_costs:
-        costs = (
-            lines_or_links.loc[dc_b, "length"]
-            * length_factor
-            * costs.at["HVDC overhead", "capital_cost"]
-        )
-    else:
-        costs = (
-            lines_or_links.loc[dc_b, "length"]
-            * length_factor
-            * (
-                (1.0 - lines_or_links.loc[dc_b, "underwater_fraction"])
-                * costs.at["HVDC overhead", "capital_cost"]
-                + lines_or_links.loc[dc_b, "underwater_fraction"]
-                * costs.at["HVDC submarine", "capital_cost"]
-            )
-            + costs.at["HVDC inverter pair", "capital_cost"]
-        )
-    lines_or_links.loc[dc_b, "capital_cost"] = costs
-
-
-def update_transmission_costs(n, costs, length_factor=1.0, simple_hvdc_costs=False):
-    n.lines["capital_cost"] = (
-        n.lines["length"] * length_factor * costs.at["HVAC overhead", "capital_cost"]
-    )
-
-    attach_dc_costs(
-        lines_or_links=n.links,
-        costs=costs,
-        length_factor=length_factor,
-        simple_hvdc_costs=simple_hvdc_costs,
-    )
-    attach_dc_costs(
-        lines_or_links=n.lines,
-        costs=costs,
-        length_factor=length_factor,
-        simple_hvdc_costs=simple_hvdc_costs,
-    )
-
-
-def calculate_renewable_capital_cost(
-    carrier: str,
-    ds: xr.Dataset,
-    costs: pd.DataFrame,
-    line_length_factor: float,
-    output_currency: str = "EUR",
-):
-    """
-    Compute the per-bus capital cost of a renewable carrier from a cost table.
-
-    For offshore wind the capital cost includes a distance-dependent grid
-    connection cost (submarine/underground), which is why the renewable profile
-    dataset ``ds`` (providing ``average_distance`` and ``underwater_fraction``)
-    is required. For all other carriers the capital cost is the carrier-level
-    value from ``costs``.
-
-    This helper is the single source of truth for renewable capital costs and is
-    used both at network build time (``attach_wind_and_solar``) and for
-    per-horizon re-costing, so the two paths can never diverge.
-
-    Parameters
-    ----------
-    carrier : str
-        Renewable carrier name (e.g. ``"onwind"``, ``"offwind-ac"``).
-    ds : xr.Dataset
-        Renewable profile dataset for the carrier.
-    costs : pd.DataFrame
-        Technology cost table indexed by technology.
-    line_length_factor : float
-        Factor scaling connection distances for offshore wind.
-    output_currency : str
-        Currency label used only for logging.
-
-    Returns
-    -------
-    capital_cost : float or pandas.Series
-        Total carrier capital cost (technology + connection). A per-bus
-        ``Series`` for offshore wind, a scalar otherwise.
-    capital_cost_tech : float
-        The technology-only component of the capital cost (a scalar per
-        carrier), i.e. excluding the distance-dependent grid connection cost.
-        For non-offshore carriers this equals ``capital_cost``. Storing this on
-        the network lets the per-horizon re-cost recover the frozen connection
-        cost as ``capital_cost - capital_cost_tech`` without needing the
-        profile geometry again.
-    """
-    supcarrier = carrier.split("-", 2)[0]
-
-    if supcarrier == "offwind":
-        capital_cost_tech = (
-            costs.at["offwind", "capital_cost"]
-            + costs.at[carrier + "-station", "capital_cost"]
-        )
-        underwater_fraction = ds["underwater_fraction"].to_pandas()
-        connection_cost = (
-            line_length_factor
-            * ds["average_distance"].to_pandas()
-            * (
-                underwater_fraction
-                * costs.at[carrier + "-connection-submarine", "capital_cost"]
-                + (1.0 - underwater_fraction)
-                * costs.at[carrier + "-connection-underground", "capital_cost"]
-            )
-        )
-        capital_cost = capital_cost_tech + connection_cost
-
-        logger.info(
-            "Added connection cost of {:0.0f}-{:0.0f} {}/MW/a to {}".format(
-                connection_cost.min(),
-                connection_cost.max(),
-                output_currency,
-                carrier,
-            )
-        )
-    else:
-        capital_cost = costs.at[carrier, "capital_cost"]
-        capital_cost_tech = capital_cost
-
-    return capital_cost, capital_cost_tech
-
-
-def update_electricity_costs(
-    n: pypsa.Network,
-    costs: pd.DataFrame,
-    renewable_carriers,
-    length_factor: float = 1.0,
-    hydro_capital_cost: bool = False,
-) -> None:
-    """
-    Re-apply all cost-vintage-dependent attributes on an already-built
-    electricity network, without rebuilding it.
-
-    This is the single entry point for per-horizon re-costing. The structural
-    network (topology, installed capacities, renewable profiles, build years,
-    and the plant-specific efficiencies/lifetimes of *existing* units) is built
-    once by ``add_electricity`` + ``add_extra_components``; this function then
-    overwrites only the economic attributes that depend on the cost year, using
-    a horizon-specific ``costs`` table. It is intended to be called at the very
-    start of ``prepare_network`` (before any opts-based cost scaling).
-
-    The formulas below mirror exactly those used at build time in
-    ``add_electricity`` and ``add_extra_components`` so the build and re-cost
-    paths cannot diverge. Attributes that do NOT depend on the cost vintage are
-    deliberately left untouched:
-      * ``p_nom``/``p_nom_min``/``p_nom_max`` and ``p_max_pu`` (capacities, profiles),
-      * ``build_year`` and the per-plant ``efficiency``/``lifetime`` of existing
-        (non-extendable) conventional generators,
-      * transmission ``efficiency`` (set from ``sector.transmission_efficiency``,
-        not from the cost table).
-
-    Offshore wind capital cost includes a distance-dependent grid connection
-    cost that is treated as frozen infrastructure across horizons. It is
-    recovered as ``capital_cost - capital_cost_tech`` (the latter stored at
-    build time, see :func:`calculate_renewable_capital_cost`), so no profile
-    geometry is needed here; only the technology component is re-costed.
-
-    Parameters
-    ----------
-    n : pypsa.Network
-        The already-built (clustered) network to re-cost in place.
-    costs : pd.DataFrame
-        Horizon-specific technology cost table indexed by technology.
-    renewable_carriers : iterable of str
-        Renewable generator carriers (used to route generator re-costing).
-    length_factor : float
-        Transmission/offshore connection length factor.
-    hydro_capital_cost : bool
-        Whether hydro reservoirs carry a capital cost (mirrors
-        ``renewable: hydro: hydro_capital_cost``).
-    """
-    renewable_carriers = set(renewable_carriers)
-
-    def _assign(df, mask, **attrs):
-        idx = df.index[mask]
-        if idx.empty:
-            return
-        for attr, value in attrs.items():
-            df.loc[idx, attr] = value
-
-    # --- Transmission lines and DC links (mirrors update_transmission_costs) ---
-    update_transmission_costs(n, costs, length_factor=length_factor)
-
-    gens = n.generators
-    storage = n.storage_units
-    stores = n.stores
-    links = n.links
-
-    # --- Generators ---
-    for carrier in gens.carrier.dropna().unique():
-        supcarrier = carrier.split("-", 2)[0]
-        is_carrier = gens.carrier == carrier
-        is_ext = is_carrier & gens.p_nom_extendable
-
-        if carrier == "ror":
-            # Run-of-river hydro: mirrors attach_hydro (no marginal_cost in
-            # the cost table for ror).
-            _assign(
-                n.generators,
-                is_carrier,
-                capital_cost=costs.at["ror", "capital_cost"],
-                efficiency=costs.at["ror", "efficiency"],
-            )
-
-        elif carrier in renewable_carriers or supcarrier in renewable_carriers:
-            # Wind/solar/csp: all cost attributes refreshed from horizon table.
-            # For offshore wind, geometry (average_distance, underwater_fraction)
-            # is stored on generators at build time; unit costs come from the
-            # new table, so both the technology and connection parts are exact.
-            if supcarrier == "offwind":
-                lifetime = costs.at["offwind", "lifetime"]
-            else:
-                lifetime = costs.at[carrier, "lifetime"]
-
-            _assign(
-                n.generators,
-                is_carrier,
-                marginal_cost=costs.at[supcarrier, "marginal_cost"],
-                efficiency=costs.at[supcarrier, "efficiency"],
-                lifetime=lifetime,
-            )
-
-            sub = gens.index[is_carrier]
-            has_geometry = (
-                "average_distance" in n.generators.columns
-                and "underwater_fraction" in n.generators.columns
-                and n.generators.loc[sub, "average_distance"].notna().all()
-            )
-
-            if supcarrier == "offwind" and has_geometry:
-                avg_dist = n.generators.loc[sub, "average_distance"]
-                uw_frac = n.generators.loc[sub, "underwater_fraction"]
-                connection_cost = (
-                    length_factor
-                    * avg_dist
-                    * (
-                        uw_frac
-                        * costs.at[carrier + "-connection-submarine", "capital_cost"]
-                        + (1.0 - uw_frac)
-                        * costs.at[carrier + "-connection-underground", "capital_cost"]
-                    )
-                )
-                n.generators.loc[sub, "capital_cost"] = (
-                    costs.at["offwind", "capital_cost"]
-                    + costs.at[carrier + "-station", "capital_cost"]
-                    + connection_cost
-                )
-            elif supcarrier == "offwind":
-                logger.warning(
-                    f"No stored geometry for offshore carrier '{carrier}'; "
-                    "capital_cost left unchanged during re-costing."
-                )
-            else:
-                n.generators.loc[sub, "capital_cost"] = costs.at[
-                    carrier, "capital_cost"
-                ]
-
-        elif carrier in costs.index:
-            # Conventional generators. marginal_cost is carrier-uniform at build
-            # time, so refresh it for all units of the carrier. capital_cost,
-            # efficiency and lifetime are only meaningful (and only refreshed)
-            # for the extendable units; existing units keep their plant-specific
-            # efficiency/lifetime and zero relevance capital_cost.
-            _assign(
-                n.generators,
-                is_carrier,
-                marginal_cost=costs.at[carrier, "marginal_cost"],
-            )
-            _assign(
-                n.generators,
-                is_ext,
-                capital_cost=costs.at[carrier, "capital_cost"],
-                efficiency=costs.at[carrier, "efficiency"],
-                lifetime=costs.at[carrier, "lifetime"],
-            )
-
-    # --- Storage units ---
-    if not storage.empty:
-        # PHS (existing): capital_cost + round-trip efficiency split (sqrt)
-        _assign(
-            n.storage_units,
-            storage.carrier == "PHS",
-            capital_cost=costs.at["PHS", "capital_cost"],
-            efficiency_store=np.sqrt(costs.at["PHS", "efficiency"]),
-            efficiency_dispatch=np.sqrt(costs.at["PHS", "efficiency"]),
-        )
-
-        # Hydro reservoir (existing)
-        _assign(
-            n.storage_units,
-            storage.carrier == "hydro",
-            capital_cost=(
-                costs.at["hydro", "capital_cost"] if hydro_capital_cost else 0.0
-            ),
-            marginal_cost=costs.at["hydro", "marginal_cost"],
-            efficiency_dispatch=costs.at["hydro", "efficiency"],
-        )
-
-        # Battery: existing (non-extendable) and extendable StorageUnits share
-        # the carrier but were costed differently at build time.
-        is_batt = storage.carrier == "battery"
-        is_batt_existing = is_batt & (~storage.p_nom_extendable)
-        is_batt_ext = is_batt & storage.p_nom_extendable
-        _assign(
-            n.storage_units,
-            is_batt_existing,
-            capital_cost=costs.at["battery", "capital_cost"],
-            marginal_cost=costs.at["battery", "marginal_cost"],
-            efficiency_store=np.sqrt(costs.at["battery", "efficiency"]),
-            efficiency_dispatch=np.sqrt(costs.at["battery", "efficiency"]),
-        )
-        _assign(
-            n.storage_units,
-            is_batt_ext,
-            capital_cost=costs.at["battery", "capital_cost"],
-            marginal_cost=costs.at["battery", "marginal_cost"],
-            efficiency_store=costs.at["battery inverter", "efficiency"],
-            efficiency_dispatch=costs.at["battery inverter", "efficiency"],
-        )
-
-        # Extendable H2 StorageUnit (only present if modelled as StorageUnit)
-        _assign(
-            n.storage_units,
-            storage.carrier == "H2",
-            capital_cost=costs.at["H2", "capital_cost"],
-            marginal_cost=costs.at["H2", "marginal_cost"],
-            efficiency_store=costs.at["electrolysis", "efficiency"],
-            efficiency_dispatch=costs.at["fuel cell", "efficiency"],
-        )
-
-    # --- Stores (Store-Link-Bus storage representation) ---
-    if not stores.empty:
-        _assign(
-            n.stores,
-            stores.carrier == "H2",
-            capital_cost=costs.at["hydrogen storage tank", "capital_cost"],
-        )
-        _assign(
-            n.stores,
-            stores.carrier == "battery",
-            capital_cost=costs.at["battery storage", "capital_cost"],
-            marginal_cost=costs.at["battery", "marginal_cost"],
-        )
-        _assign(
-            n.stores,
-            stores.carrier == "csp",
-            capital_cost=costs.at["csp-tower TES", "capital_cost"],
-            marginal_cost=costs.at["csp-tower TES", "marginal_cost"],
-        )
-
-    # --- Extendable links (charging/discharging, pipelines, csp) ---
-    if not links.empty:
-        _assign(
-            n.links,
-            links.carrier == "H2 electrolysis",
-            efficiency=costs.at["electrolysis", "efficiency"],
-            capital_cost=costs.at["electrolysis", "capital_cost"],
-            marginal_cost=costs.at["electrolysis", "marginal_cost"],
-        )
-        _assign(
-            n.links,
-            links.carrier == "H2 fuel cell",
-            efficiency=costs.at["fuel cell", "efficiency"],
-            capital_cost=costs.at["fuel cell", "capital_cost"]
-            * costs.at["fuel cell", "efficiency"],
-            marginal_cost=costs.at["fuel cell", "marginal_cost"],
-        )
-        _assign(
-            n.links,
-            links.carrier == "battery charger",
-            efficiency=costs.at["battery inverter", "efficiency"],
-            capital_cost=costs.at["battery inverter", "capital_cost"],
-            marginal_cost=costs.at["battery inverter", "marginal_cost"],
-        )
-        _assign(
-            n.links,
-            links.carrier == "battery discharger",
-            efficiency=costs.at["battery inverter", "efficiency"],
-            marginal_cost=costs.at["battery inverter", "marginal_cost"],
-        )
-        _assign(
-            n.links,
-            links.carrier == "csp",
-            efficiency=costs.at["csp-tower", "efficiency"],
-            capital_cost=costs.at["csp-tower", "capital_cost"],
-            marginal_cost=costs.at["csp-tower", "marginal_cost"],
-        )
-        # H2 pipelines: capital_cost scales with length; efficiency is set from
-        # the transmission-efficiency config (not the cost table) so it is left
-        # untouched here.
-        h2_pipe = links.carrier.str.startswith("H2 pipeline", na=False)
-        if h2_pipe.any():
-            n.links.loc[h2_pipe, "capital_cost"] = (
-                costs.at["H2 pipeline", "capital_cost"] * n.links.loc[h2_pipe, "length"]
-            )
 
 
 def get_grouping_year(build_year: int, grouping_years: list) -> int:
